Refactor `eslint.config.js` to update global variables for GAS compatibility; replace `99_BaseTaskArtifact.js` with `0_BaseTaskArtifact.js` in multiple artifact files and remove the obsolete `99_BaseTaskArtifact.js`.
