+++ eslint.config.js
@@ -6,15 +6,31 @@ module.exports = [
       ecmaVersion: 'latest',
       sourceType: 'script',
       globals: {
-        // --- Add your custom globals here ---
-        ClassroomApiClient: 'readonly',
-        DriveManager: 'readonly',
+        // Project singletons and global classes exposed to GAS
         Assessment: 'readonly',
         Assignment: 'readonly',
-        GoogleClassroom: 'readonly',
+        AssessmentRecord: 'readonly',
         Student: 'readonly',
-        StudentTask: 'readonly',
-        Task: 'readonly',
+        StudentSubmission: 'readonly',
+        TaskDefinition: 'readonly',
+        BaseTaskArtifact: 'readonly',
+        TextTaskArtifact: 'readonly',
+        TableTaskArtifact: 'readonly',
+        SpreadsheetTaskArtifact: 'readonly',
+        ImageTaskArtifact: 'readonly',
+        ArtifactFactory: 'readonly',
+
+        // Managers / controllers / utilities used as globals or singletons
+        Utils: 'readonly',
+        ProgressTracker: 'readonly',
+        UIManager: 'readonly',
+        ConfigurationManagerClass: 'readonly',
+        LLMRequestManager: 'readonly',
+        CacheManager: 'readonly',
+        ImageManager: 'readonly',
+        TriggerController: 'readonly',
+
+        // Sheet managers and extractors
         BaseSheetManager: 'readonly',
         AnalysisSheetManager: 'readonly',
         ClassAssessmentSheet: 'readonly',
@@ -24,15 +40,6 @@ module.exports = [
         OverviewSheetManager: 'readonly',
         SummarySheetManager: 'readonly',
         AssignmentPropertiesManager: 'readonly',
-        ConfigurationManagerClass: 'readonly',
-        ProgressTracker: 'readonly',
-        TriggerController: 'readonly',
-        Utils: 'readonly',
-        CacheManager: 'readonly',
-        ImageManager: 'readonly',
-        LLMRequestManager: 'readonly',
-        UIManager: 'readonly',
-        MainController: 'readonly',
       },
     },
     plugins: { googleappsscript },
@@ -44,37 +51,44 @@ module.exports = [
           paths: [
             {
               name: '../src/AdminSheet/Models/Artifacts/0_BaseTaskArtifact.js',
-              message: 'Import from Models/Artifacts/index.js instead of individual artifact files.'
+              message:
+                'Import from Models/Artifacts/index.js instead of individual artifact files.',
             },
             {
               name: '../src/AdminSheet/Models/Artifacts/1_TextTaskArtifact.js',
-              message: 'Import from Models/Artifacts/index.js instead of individual artifact files.'
+              message:
+                'Import from Models/Artifacts/index.js instead of individual artifact files.',
             },
             {
               name: '../src/AdminSheet/Models/Artifacts/2_TableTaskArtifact.js',
-              message: 'Import from Models/Artifacts/index.js instead of individual artifact files.'
+              message:
+                'Import from Models/Artifacts/index.js instead of individual artifact files.',
             },
             {
               name: '../src/AdminSheet/Models/Artifacts/3_SpreadsheetTaskArtifact.js',
-              message: 'Import from Models/Artifacts/index.js instead of individual artifact files.'
+              message:
+                'Import from Models/Artifacts/index.js instead of individual artifact files.',
             },
             {
               name: '../src/AdminSheet/Models/Artifacts/4_ImageTaskArtifact.js',
-              message: 'Import from Models/Artifacts/index.js instead of individual artifact files.'
+              message:
+                'Import from Models/Artifacts/index.js instead of individual artifact files.',
             },
             {
               name: '../src/AdminSheet/Models/Artifacts/5_ArtifactFactory.js',
-              message: 'Import from Models/Artifacts/index.js instead of individual artifact files.'
-            }
+              message:
+                'Import from Models/Artifacts/index.js instead of individual artifact files.',
+            },
           ],
           patterns: [
             {
               group: ['**/Models/Artifacts/[0-9]_*.js'],
-              message: 'Use Models/Artifacts/index.js as the import surface, not numbered artifact files.'
-            }
-          ]
-        }
-      ]
-    }
+              message:
+                'Use Models/Artifacts/index.js as the import surface, not numbered artifact files.',
+            },
+          ],
+        },
+      ],
+    },
   },
 ];
