+++ src/AdminSheet/Models/Artifacts/3_SpreadsheetTaskArtifact.js
@@ -1,6 +1,6 @@
 const BaseTaskArtifact =
   typeof module !== 'undefined' && module.exports
-    ? require('./99_BaseTaskArtifact.js')
+    ? require('./0_BaseTaskArtifact.js')
     : this.BaseTaskArtifact;
 
 class SpreadsheetTaskArtifact extends BaseTaskArtifact {
