+++ src/AdminSheet/Models/Artifacts/4_ImageTaskArtifact.js
@@ -1,6 +1,6 @@
 const BaseTaskArtifact =
   typeof module !== 'undefined' && module.exports
-    ? require('./99_BaseTaskArtifact.js')
+    ? require('./0_BaseTaskArtifact.js')
     : this.BaseTaskArtifact;
 
 class ImageTaskArtifact extends BaseTaskArtifact {
